Moved PasswordEncryptTest
